fix TypeError + ValueError missing object in array

+++ FitFunctions/fit_functions.py
@@ -2,4 +2,4 @@ import numpy as np
 
 
 def sin(x, a, b, c, d):
-    return a * np.sin((b * x) + c) + d
+    return a * np.sin((b * x) + c, dtype=np.longdouble) + d
